Add Dockerfile
For now, this allows us to build a container image using dotnet core 2.2
on Linux.

We have to install 'libgdiplus' and 'libc6-devel' as runtime
dependencies, otherwise we won't be able to generate images.

Because we're on Docker, we also need to adjust the launchSettings to
listen on any address, not only 'localhost'. From inside the Docker
container it looks like we're connecting from remote.

Eventually, I'd like to build this for multiple architectures with
buildx. However, I think the first step will be to build it for arm64,
so I can run it on my Raspberry Pi.