Migrate to .Net Core 3.1
We no longer need to include the AspNet packages, because these are now
part of the .Net Core SDK.

We also have to replace 2.1 with 3.1 in the csproj file.

When building, I noticed that AspNetCore 3.1 seems to disallow
synchronous method calls, so I had to convert the Save() method to an
async method SaveAsync(). To keep the existing API, I added a
synchronous wrapper method Save().