Migrate to .Net 6.0
I'm not sure why this worked. I've been playing around with the 6.0 SDK
on my machine the whole day, but now after the incremental updates, it
no longer seems to care about System.Drawing.Common. At least not, when
I build it with Docker.

I just confirmed that I'm still having the same issue on my local
machine. I'm wary of releasing this, but the Docker image seems to be
fine.