Rescale image to fit text and properly center all elements
The issue reported by Kendall is caused by multiple interferring
calculations.

First of all, the image width is mainly determined by the chord box and
**not** by the chord name, even though the latter can be much longer /
wider. This is solved by resizing the image. The solution is not ideal
and produces some redundant code. The main issue is that we have to
initialize a Graphics instance first to measure the string. But to
initialize the Graphics, we need a Bitmap instance of a certain width.

Another problem I noticed is that neither the chord box, nor the chord
name were properly centered. The widths of both are calculated, but
until now they weren't used to place the elements properly. This is now
fixed so that no matter the final width of the image, the chord box and
name stay nicely centered. I'll run this by Kendall to make sure that
centering the text is what he would expect as well.

This relates to #4.