Enable full barre option for fingers that only occupy one position
Usually we only draw barre bars whenever a finger is used for two
strings in the same position. However, the full barre is a bit
different. It should signal that the finger is used to bar all strings.
If the rest of the strings are pressed with other fingers, we should
still draw a bar.

This will make the option behave weirdly whenever open strings are found
in the middle, but the option can simply be deactivated in those cases.