Hardcode 'perc' value to scale fonts
I got this from spinning up a container of the old version and logging
the ratio it returned. It should be a static value and as far as I can
see, there's nothing similar in ImageSharp to reproduce the calculation.

While at it, I noticed that the fontsize for the finger positions at the
bottom doesn't scale with the rest. I experimented a bit and IMO it
makes sense to scale this as well. The result looks much nicer and
produces more consistent results when changing the size.