Fix chordname position and size
This gets rid of the insanely complicated positioning algorithm for the
chordname. I don't really care if it's too long, it will always make
sense to put it dead in the center. Apart from that, we also grow the
canvas size to always fit the chordname.

I had to add an offset compared to the old implementation. I don't fully
understand why that's necessary, or rather why it wasn't necessary in
the past, but it now looks to be at almost the same height as before.

I also had to introduce some additional spacing between the letters,
which the old implementation didn't need. Without this additional space,
letters (especially superscript) would end up overlapping each other.