Downgrade to .NET 6.0
There's currently a problem with building .NET 8 either with QEMU or a
remote builder. While the QEMU issue is documented and Microsoft simply
refuses to look into and fix it, the other issue is not very well
understood. When I use my Raspberry Pi as a remote builder, I run into
timeouts when querying the NuGet repository.

Doing the same with 'curl' works perfectly fine. I've tried a bunch of
environment variables, but to no avail. I'm running out of patience and
that's why I'm reverting back to .NET 6.0.

This reverts commit 29ff7d0f2dac7d5022d248e2e86718540548e8f5.