Add intermediary sizes
Kendall Miller proposed adding new sizes. They were particularly
interested in a size between 1 and 2, because when inserted into a tab
this size would match up nicely with commercially sold tabs.

To keep the original sizes intact, I increased the maximum to a size of
20 (which would be 10.5 in the old scale). There's not much sense in
going lower than size 1, so that's still the lower limit.