Remove character limit on input fields
I'm working on adding a new feature to draw multiple fingers per fret,
e.g. to show that a barre starts at a certain fret, although another
finger is in a higher position on the same fret.

My idea is to allow inputs like "2/4" for fret positions and "1/3" for
finger positions, indicating that the index finger is on the 2nd fret
and the ring finger is on the 4th fret, but both are on the same string.