Allow multiple fingers on the same string
This allows drawing a highly accurate and specific chord, where we can
show exactly where a barre starts, even if the starting point is on a
string that it occupied by another finger in a higher position.

I opted to solve this by allowing slashes for both fret positions and
fingers. These are translated to '+' characters in the final query
string. I chose this for the following reasons.

1. The ampersand & would interfere with the query parameters in the URL.
2. We are already using '-' if we have fret positions that require
   double digits.