Fix the issue with the missing plus sign
We end up with spaces in the fingers parameter, but we need plus
characters. It took me a while until I figured out that it's because of
the URL sanitization. I'm not particularly proud of this hack, as it
basically replaces all the sanitized spaces with plus characters, but
it's a pragmatic and easy solution.