Remove nested elements in index.html
So I don't understand what's going on AT ALL!

As soon as I added `<p>` or `<li>` tags to the elements, the links would
no longer reload the site. The URL would still change and forcing a
refresh with `CTRL+F5` would still get you to the right page and image,
but it no longer worked by clicking the link.

I have absolutely no idea why this is the case, but I'm done looking for
the answer. So I just got rid of all the nested elements and decided to
try my best at formatting things this way. I guess it looks okay enough.